tabela produtos

+++ Curso PHP/Oficina/painel-adm/categorias/excluir.php
@@ -3,9 +3,19 @@ require_once("../../conexao.php");
 
 $id = $_POST['id'];
 
+//EXCLUIR SOMENTE SE NÃO TIVER REGISTROS RELACIONADOS
+$query_tot = $pdo->query("SELECT * FROM produtos where categoria = '$id'");
+$res_tot = $query_tot->fetchAll(PDO::FETCH_ASSOC);
+$total_produtos = @count($res_tot);
 
-$pdo->query("DELETE FROM categorias WHERE id = '$id'");
+if($total_produtos == 0){
+	$pdo->query("DELETE FROM categorias WHERE id = '$id'");
+	echo 'Excluído com Sucesso!';
+}else{
+	echo 'Você possui produtos relacionados a esta categoria, primeiro exclua os produtos para depois excluir a categoria!';
+}
 
-echo 'Excluído com Sucesso!';
+
+
 
 ?>
